Fix the math

+++ Dynamixel_sample_controller/Routine.cpp
@@ -12,7 +12,7 @@
   #define DEBUG_SERIAL Serial
 #endif
 
-void peristalsisRoutine (DynamixelShield &dxl, int8_t worm_pattern[][SEGMENT_NUMBER], int number_Of_Motor, int32_t calibration[], uint8_t DXL_ID[], int iteration, int32_t full_contraction, bool not_pause, float turningrate){
+void peristalsisRoutine (DynamixelShield &dxl, int8_t worm_pattern[][SEGMENT_NUMBER], int number_Of_Motor, int32_t calibration[], uint8_t DXL_ID[], int iteration, int32_t full_contraction, bool not_pause, double turningrate){
 
   for(int i = 0;i<number_Of_Motor/2;i++){
 	    int32_t increase_amount =  int32_t(not_pause) * (50 + full_contraction * int32_t(worm_pattern[iteration][i])); //add 30 offset to remove slack cables
@@ -41,7 +41,7 @@ void peristalsisRoutine (DynamixelShield &dxl, int8_t worm_pattern[][SEGMENT_NUM
    //delay(200);
 }
 
-void undulationRoutine (DynamixelShield &dxl, int8_t worm_pattern_turning[][SEGMENT_NUMBER], int number_Of_Motor, int32_t calibration[], uint8_t DXL_ID[], int iteration, int32_t full_contraction, bool not_pause, float turningrate){
+void undulationRoutine (DynamixelShield &dxl, int8_t worm_pattern_turning[][SEGMENT_NUMBER], int number_Of_Motor, int32_t calibration[], uint8_t DXL_ID[], int iteration, int32_t full_contraction, bool not_pause, double turningrate){
 
   for(int i = 0;i<number_Of_Motor/2;i++){
       int32_t relax = abs(worm_pattern_turning[iteration][i]);
